requirements.txt
# Core — install PyTorch with CUDA from https://pytorch.org/get-started/
# pip install torch torchvision --index-url https://download.pytorch.org/whl/cu130
torch>=2.10.0
torchvision>=0.25.0
numpy==1.26.4

# Wan2.1 model
diffusers==0.37.0
transformers==5.3.0
tokenizers==0.22.2
accelerate==1.13.0
einops==0.8.0
easydict==1.13
ftfy==6.3.1
regex==2026.2.19

# Image / Video I/O
Pillow==11.0.0
opencv-python==4.10.0.84
imageio==2.34.2
imageio-ffmpeg==0.5.1

# Compression
compressai==1.2.8

# Metrics
pytorch-msssim==1.0.0
lpips==0.1.4

# Utilities
tqdm==4.67.3
